Common Tiler features tests are now in the same file
New file 'test_tiler.py' containing the tests for the Tiler class features (scale, offset, crs, loa, lod1, obj)
