+++ tests/tiler_test_data/loa_polygons/polygons_lod_tree_test.geojson
